+++ nodeconductor/structure/serializers.py
@@ -584,10 +584,9 @@ class CustomerPermissionSerializer(PermissionFieldFilteringMixin,
     def validate(self, data):
         customer = data['customer']
         user = data['user']
-        role = data['role']
 
-        if customer.has_user(user, role):
-            raise serializers.ValidationError('The fields customer, user, role must make a unique set.')
+        if customer.has_user(user):
+            raise serializers.ValidationError('The fields customer and user must make a unique set.')
 
         return data
 
@@ -658,10 +657,9 @@ class ProjectPermissionSerializer(PermissionFieldFilteringMixin,
     def validate(self, data):
         project = data['project']
         user = data['user']
-        role = data['role']
 
-        if project.has_user(user, role):
-            raise serializers.ValidationError('The fields project, user, role must make a unique set.')
+        if project.has_user(user):
+            raise serializers.ValidationError('The fields project and user must make a unique set.')
 
         return data
 
@@ -1337,6 +1335,10 @@ class BaseResourceSerializer(six.with_metaclass(ResourceSerializerMetaclass,
         read_only=True,
         lookup_field='uuid')
     service_settings_uuid = serializers.ReadOnlyField(source='service_project_link.service.settings.uuid')
+    service_settings_state = serializers.ReadOnlyField(
+        source='service_project_link.service.settings.human_readable_state')
+    service_settings_error_message = serializers.ReadOnlyField(
+        source='service_project_link.service.settings.error_message')
 
     customer = serializers.HyperlinkedRelatedField(
         source='service_project_link.project.customer',
@@ -1362,6 +1364,7 @@ class BaseResourceSerializer(six.with_metaclass(ResourceSerializerMetaclass,
             'url', 'uuid', 'name', 'description', 'start_time',
             'service', 'service_name', 'service_uuid',
             'service_settings', 'service_settings_uuid',
+            'service_settings_state', 'service_settings_error_message',
             'project', 'project_name', 'project_uuid',
             'customer', 'customer_name', 'customer_native_name', 'customer_abbreviation',
             'project_groups', 'tags', 'error_message',